# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grooming CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /financial-entries:
    get:
      operationId: listFinancialEntries
      tags: [financial]
      summary: List financial entries
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of financial entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialEntry"
    post:
      operationId: createFinancialEntry
      tags: [financial]
      summary: Create a financial entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialEntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialEntry"
  /financial-entries/summary:
    get:
      operationId: getFinancialSummary
      tags: [financial]
      summary: Get financial summary (totals by type and period)
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"
  /financial-entries/{id}:
    get:
      operationId: getFinancialEntry
      tags: [financial]
      summary: Get a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financial entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialEntry"
    patch:
      operationId: updateFinancialEntry
      tags: [financial]
      summary: Update a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialEntryUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialEntry"
    delete:
      operationId: deleteFinancialEntry
      tags: [financial]
      summary: Delete a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Message Templates
  /message-templates:
    get:
      operationId: listMessageTemplates
      tags: [templates]
      summary: List message templates
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageTemplate"
    post:
      operationId: createMessageTemplate
      tags: [templates]
      summary: Create a message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
  /message-templates/{id}:
    get:
      operationId: getMessageTemplate
      tags: [templates]
      summary: Get a message template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
    patch:
      operationId: updateMessageTemplate
      tags: [templates]
      summary: Update a message template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageTemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
    delete:
      operationId: deleteMessageTemplate
      tags: [templates]
      summary: Delete a message template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Leads
...
components:
...
  schemas:
...
    FinancialEntry:
      type: object
      required: [id, tenantId, type, description, amount, date, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        type:
          type: string
          enum: [receita, despesa, despesa_fixa]
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        category:
          type: ["string", "null"]
        appointmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    FinancialEntryInput:
      type: object
      required: [tenantId, type, description, amount, date]
      properties:
        tenantId:
          type: integer
        type:
          type: string
          enum: [receita, despesa, despesa_fixa]
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        category:
          type: string
        appointmentId:
          type: integer
    FinancialEntryUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [receita, despesa, despesa_fixa]
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        category:
          type: string
    FinancialSummary:
      type: object
      required: [totalReceitas, totalDespesas, totalDespesasFixas, saldo]
      properties:
        totalReceitas:
          type: number
        totalDespesas:
          type: number
        totalDespesasFixas:
          type: number
        saldo:
          type: number
        byMonth:
          type: array
          items:
            type: object
            required: [month, receitas, despesas, saldo]
            properties:
              month:
                type: string
              receitas:
                type: number
              despesas:
                type: number
              saldo:
                type: number

    # Message Template
    MessageTemplate:
      type: object
      required: [id, tenantId, name, type, content, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [confirmacao, lembrete, leads, agradecimento]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageTemplateInput:
      type: object
      required: [tenantId, name, type, content]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [confirmacao, lembrete, leads, agradecimento]
        content:
          type: string
    MessageTemplateUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [confirmacao, lembrete, leads, agradecimento]
        content:
          type: string

    # Lead
...